@@ -1,223 +1,208 @@
-﻿using System.Collections.Generic;
-using Moq;
+﻿using System;
+using System.Collections.Generic;
 using NUnit.Framework;
 using SourcemapToolkit.SourcemapParser;
+using SourcemapTools.CallstackDeminifier.Internal;
 
-namespace SourcemapToolkit.CallstackDeminifier.UnitTests
+namespace SourcemapToolkit.CallstackDeminifier.UnitTests;
+
+public class StackFrameDeminifierUnitTests
 {
-
-	public class StackFrameDeminifierUnitTests
+	private static IStackFrameDeminifier GetStackFrameDeminifierWithMockDependencies(ISourceMapStore? sourceMapStore = null, IFunctionMapStore? functionMapStore = null, IFunctionMapConsumer? functionMapConsumer = null, bool useSimpleStackFrameDeminifier = false)
 	{
-		private static IStackFrameDeminifier GetStackFrameDeminifierWithMockDependencies(ISourceMapStore? sourceMapStore = null, IFunctionMapStore? functionMapStore = null, IFunctionMapConsumer? functionMapConsumer = null, bool useSimpleStackFrameDeminier = false)
-		{
-			if (sourceMapStore == null)
-			{
-				sourceMapStore = new Mock<ISourceMapStore>().Object;
-			}
+		sourceMapStore ??= new ISourceMapStoreMock(_ => null);
 
-			if (functionMapStore == null)
-			{
-				functionMapStore = new Mock<IFunctionMapStore>().Object;
-			}
+		functionMapStore ??= new IFunctionMapStoreMock(_ => null);
 
-			if (functionMapConsumer == null)
-			{
-				functionMapConsumer = new Mock<IFunctionMapConsumer>().Object;
-			}
+		functionMapConsumer ??= new IFunctionMapConsumerMock((x, y) => null);
 
-			return useSimpleStackFrameDeminier
-				? new MethodNameStackFrameDeminifier(functionMapStore, functionMapConsumer)
-				: new StackFrameDeminifier(sourceMapStore, functionMapStore, functionMapConsumer);
-		}
-
-		[Test]
-		public void DeminifyStackFrame_StackFrameNullProperties_DoesNotThrowException([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var stackFrame = new StackFrame(null);
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies();
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-		[Test]
-		public void SimpleStackFrameDeminierDeminifyStackFrame_FunctionMapReturnsNull_NoFunctionMapDeminificationError([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var filePath = "foo";
-			var stackFrame = new StackFrame(null) { FilePath = filePath };
-			var functionMapStore = new Mock<IFunctionMapStore>();
-			functionMapStore.Setup(c => c.GetFunctionMapForSourceCode(filePath))
-				.Returns<FunctionMapEntry>(null);
-
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore.Object, useSimpleStackFrameDeminier: true);
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.AreEqual(DeminificationError.NoSourceCodeProvided, stackFrameDeminification.DeminificationError);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-		[Test]
-		public void SimpleStackFrameDeminierDeminifyStackFrame_GetWRappingFunctionForSourceLocationReturnsNull_NoWrapingFunctionDeminificationError([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var filePath = "foo";
-			var stackFrame = new StackFrame(null) { FilePath = filePath };
-			var functionMapStore = new Mock<IFunctionMapStore>();
-			functionMapStore.Setup(c => c.GetFunctionMapForSourceCode(filePath))
-				.Returns(new List<FunctionMapEntry>());
-			var functionMapConsumer = new Mock<IFunctionMapConsumer>();
-			functionMapConsumer.Setup(c => c.GetWrappingFunctionForSourceLocation(It.IsAny<SourcePosition>(), It.IsAny<List<FunctionMapEntry>>()))
-				.Returns<FunctionMapEntry?>(null);
-
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore.Object, functionMapConsumer: functionMapConsumer.Object, useSimpleStackFrameDeminier: true);
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.AreEqual(DeminificationError.NoWrapingFunctionFound, stackFrameDeminification.DeminificationError);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-		[Test]
-		public void SimpleStackFrameDeminierDeminifyStackFrame_WrapingFunctionFound_NoDeminificationError([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var filePath = "foo";
-			var wrapingFunctionMapEntry = CreateFunctionMapEntry("DeminifiedFoo");
-			var stackFrame = new StackFrame(null) { FilePath = filePath };
-			var functionMapStore = new Mock<IFunctionMapStore>();
-			functionMapStore.Setup(c => c.GetFunctionMapForSourceCode(filePath))
-				.Returns(new List<FunctionMapEntry>());
-			var functionMapConsumer = new Mock<IFunctionMapConsumer>();
-			functionMapConsumer.Setup(c => c.GetWrappingFunctionForSourceLocation(It.IsAny<SourcePosition>(), It.IsAny<List<FunctionMapEntry>>()))
-				.Returns(wrapingFunctionMapEntry);
-
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore.Object, functionMapConsumer: functionMapConsumer.Object, useSimpleStackFrameDeminier: true);
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.AreEqual(DeminificationError.None, stackFrameDeminification.DeminificationError);
-			Assert.AreEqual(wrapingFunctionMapEntry.DeminifiedMethodName, stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-
-		[Test]
-		public void StackFrameDeminierDeminifyStackFrame_SourceMapProviderReturnsNull_NoSourcemapProvidedError([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var filePath = "foo";
-			var wrapingFunctionMapEntry = CreateFunctionMapEntry(deminifiedMethodName: "DeminifiedFoo");
-			var stackFrame = new StackFrame(null) { FilePath = filePath };
-			var functionMapStore = new Mock<IFunctionMapStore>();
-			functionMapStore.Setup(c => c.GetFunctionMapForSourceCode(filePath))
-				.Returns(new List<FunctionMapEntry>());
-			var functionMapConsumer = new Mock<IFunctionMapConsumer>();
-			functionMapConsumer.Setup(c => c.GetWrappingFunctionForSourceLocation(It.IsAny<SourcePosition>(), It.IsAny<List<FunctionMapEntry>>()))
-				.Returns(wrapingFunctionMapEntry);
-
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore.Object, functionMapConsumer: functionMapConsumer.Object);
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.AreEqual(DeminificationError.NoSourceMap, stackFrameDeminification.DeminificationError);
-			Assert.AreEqual(wrapingFunctionMapEntry.DeminifiedMethodName, stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-		[Test]
-		public void StackFrameDeminierDeminifyStackFrame_SourceMapParsingNull_SourceMapFailedToParseError([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var filePath = "foo";
-			var wrapingFunctionMapEntry = CreateFunctionMapEntry(deminifiedMethodName: "DeminifiedFoo");
-			var stackFrame = new StackFrame(null) { FilePath = filePath };
-			var functionMapStore = new Mock<IFunctionMapStore>();
-			functionMapStore.Setup(c => c.GetFunctionMapForSourceCode(filePath))
-				.Returns(new List<FunctionMapEntry>());
-			var functionMapConsumer = new Mock<IFunctionMapConsumer>();
-			functionMapConsumer.Setup(c => c.GetWrappingFunctionForSourceLocation(It.IsAny<SourcePosition>(), It.IsAny<List<FunctionMapEntry>>()))
-				.Returns(wrapingFunctionMapEntry);
-			var sourceMapStore = new Mock<ISourceMapStore>();
-			sourceMapStore.Setup(c => c.GetSourceMapForUrl(It.IsAny<string>())).Returns(CreateSourceMap());
-
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(sourceMapStore: sourceMapStore.Object, functionMapStore: functionMapStore.Object, functionMapConsumer: functionMapConsumer.Object);
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.AreEqual(DeminificationError.NoMatchingMapingInSourceMap, stackFrameDeminification.DeminificationError);
-			Assert.AreEqual(wrapingFunctionMapEntry.DeminifiedMethodName, stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-		[Test]
-		public void StackFrameDeminierDeminifyStackFrame_SourceMapGeneratedMappingEntryNull_NoMatchingMapingInSourceMapError([Values] bool preferSourceMapsSymbols)
-		{
-			// Arrange
-			var filePath = "foo";
-			var wrapingFunctionMapEntry = CreateFunctionMapEntry(deminifiedMethodName: "DeminifiedFoo");
-			var stackFrame = new StackFrame(null) { FilePath = filePath };
-			var functionMapStore = new Mock<IFunctionMapStore>();
-			functionMapStore.Setup(c => c.GetFunctionMapForSourceCode(filePath))
-				.Returns(new List<FunctionMapEntry>());
-			var sourceMapStore = new Mock<ISourceMapStore>();
-			var sourceMap = CreateSourceMap(new List<MappingEntry>());
-
-			sourceMapStore.Setup(c => c.GetSourceMapForUrl(It.IsAny<string>())).Returns(sourceMap);
-			var functionMapConsumer = new Mock<IFunctionMapConsumer>();
-			functionMapConsumer.Setup(c => c.GetWrappingFunctionForSourceLocation(It.IsAny<SourcePosition>(), It.IsAny<List<FunctionMapEntry>>()))
-				.Returns(wrapingFunctionMapEntry);
-
-			var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(sourceMapStore: sourceMapStore.Object, functionMapStore: functionMapStore.Object, functionMapConsumer: functionMapConsumer.Object);
-
-			// Act
-			var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
-
-			// Assert
-			Assert.AreEqual(DeminificationError.NoMatchingMapingInSourceMap, stackFrameDeminification.DeminificationError);
-			Assert.AreEqual(wrapingFunctionMapEntry.DeminifiedMethodName, stackFrameDeminification.DeminifiedStackFrame.MethodName);
-			Assert.AreEqual(SourcePosition.NotFound, stackFrameDeminification.DeminifiedStackFrame.SourcePosition);
-			Assert.Null(stackFrameDeminification.DeminifiedStackFrame.FilePath);
-		}
-
-
-		private static FunctionMapEntry CreateFunctionMapEntry(string deminifiedMethodName) => new(
-			default!,
-			deminifiedMethodName,
-			default,
-			default);
-
-		private static SourceMap CreateSourceMap(List<MappingEntry>? parsedMappings = null) => new(
-			default,
-			default,
-			default,
-			default,
-			default,
-			parsedMappings ?? new List<MappingEntry>(),
-			default);
+		return useSimpleStackFrameDeminifier
+			? new MethodNameStackFrameDeminifier(functionMapStore, functionMapConsumer)
+			: new StackFrameDeminifier(sourceMapStore, functionMapStore, functionMapConsumer);
 	}
+
+	[Test]
+	public void DeminifyStackFrame_StackFrameNullProperties_DoesNotThrowException([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var stackFrame = new StackFrame(null);
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies();
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.Null);
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+		});
+	}
+
+	[Test]
+	public void SimpleStackFrameDeminifierDeminifyStackFrame_FunctionMapReturnsNull_NoFunctionMapDeminificationError([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var filePath = "foo";
+		var stackFrame = new StackFrame(null) { FilePath = filePath };
+		var functionMapStore = new IFunctionMapStoreMock(_ => null);
+
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore, useSimpleStackFrameDeminifier: true);
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminificationError, Is.EqualTo(DeminificationError.NoSourceCodeProvided));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.Null);
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+		});
+	}
+
+	[Test]
+	public void SimpleStackFrameDeminifierDeminifyStackFrame_GetWRappingFunctionForSourceLocationReturnsNull_NoWrappingFunctionDeminificationError([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var filePath = "foo";
+		var stackFrame = new StackFrame(null) { FilePath = filePath };
+		var functionMapStore = new IFunctionMapStoreMock(x => x == filePath ? new List<FunctionMapEntry>() : null);
+		var functionMapConsumer = new IFunctionMapConsumerMock((x, y) => null);
+
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore, functionMapConsumer: functionMapConsumer, useSimpleStackFrameDeminifier: true);
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminificationError, Is.EqualTo(DeminificationError.NoWrappingFunctionFound));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.Null);
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+		});
+	}
+
+	[Test]
+	public void SimpleStackFrameDeminifierDeminifyStackFrame_WrappingFunctionFound_NoDeminificationError([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var filePath = "foo";
+		var wrappingFunctionMapEntry = CreateFunctionMapEntry("DeminifiedFoo");
+		var stackFrame = new StackFrame(null) { FilePath = filePath };
+		var functionMapStore = new IFunctionMapStoreMock(x => x == filePath ? new List<FunctionMapEntry>() : null);
+		var functionMapConsumer = new IFunctionMapConsumerMock((x, y) => wrappingFunctionMapEntry);
+
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore, functionMapConsumer: functionMapConsumer, useSimpleStackFrameDeminifier: true);
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminificationError, Is.EqualTo(DeminificationError.None));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.EqualTo(wrappingFunctionMapEntry.DeminifiedMethodName));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+		});
+		Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+	}
+
+	[Test]
+	public void StackFrameDeminifierDeminifyStackFrame_SourceMapProviderReturnsNull_NoSourcemapProvidedError([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var filePath = "foo";
+		var wrappingFunctionMapEntry = CreateFunctionMapEntry(deminifiedMethodName: "DeminifiedFoo");
+		var stackFrame = new StackFrame(null) { FilePath = filePath };
+		var functionMapStore = new IFunctionMapStoreMock(x => x == filePath ? new List<FunctionMapEntry>() : null);
+		var functionMapConsumer = new IFunctionMapConsumerMock((x, y) => wrappingFunctionMapEntry);
+
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(functionMapStore: functionMapStore, functionMapConsumer: functionMapConsumer);
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminificationError, Is.EqualTo(DeminificationError.NoSourceMap));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.EqualTo(wrappingFunctionMapEntry.DeminifiedMethodName));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+		});
+		Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+	}
+
+	[Test]
+	public void StackFrameDeminifierDeminifyStackFrame_SourceMapParsingNull_SourceMapFailedToParseError([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var filePath = "foo";
+		var wrappingFunctionMapEntry = CreateFunctionMapEntry(deminifiedMethodName: "DeminifiedFoo");
+		var stackFrame = new StackFrame(null) { FilePath = filePath };
+		var functionMapStore = new IFunctionMapStoreMock(x => x == filePath ? new List<FunctionMapEntry>() : null);
+		var functionMapConsumer = new IFunctionMapConsumerMock((x, y) => wrappingFunctionMapEntry);
+		var sourceMapStore = new ISourceMapStoreMock(_ => CreateSourceMap());
+
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(sourceMapStore: sourceMapStore, functionMapStore: functionMapStore, functionMapConsumer: functionMapConsumer);
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminificationError, Is.EqualTo(DeminificationError.NoMatchingMappingInSourceMap));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.EqualTo(wrappingFunctionMapEntry.DeminifiedMethodName));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+		});
+		Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+	}
+
+	[Test]
+	public void StackFrameDeminifierDeminifyStackFrame_SourceMapGeneratedMappingEntryNull_NoMatchingMappingInSourceMapError([Values] bool preferSourceMapsSymbols)
+	{
+		// Arrange
+		var filePath = "foo";
+		var wrappingFunctionMapEntry = CreateFunctionMapEntry(deminifiedMethodName: "DeminifiedFoo");
+		var stackFrame = new StackFrame(null) { FilePath = filePath };
+		var functionMapStore = new IFunctionMapStoreMock(x => x == filePath ? new List<FunctionMapEntry>() : null);
+		var sourceMap = CreateSourceMap([]);
+		var sourceMapStore = new ISourceMapStoreMock(_ => sourceMap);
+
+		var functionMapConsumer = new IFunctionMapConsumerMock((x, y) => wrappingFunctionMapEntry);
+
+		var stackFrameDeminifier = GetStackFrameDeminifierWithMockDependencies(sourceMapStore: sourceMapStore, functionMapStore: functionMapStore, functionMapConsumer: functionMapConsumer);
+
+		// Act
+		var stackFrameDeminification = stackFrameDeminifier.DeminifyStackFrame(stackFrame, null, preferSourceMapsSymbols);
+
+		Assert.Multiple(() =>
+		{
+			// Assert
+			Assert.That(stackFrameDeminification.DeminificationError, Is.EqualTo(DeminificationError.NoMatchingMappingInSourceMap));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.MethodName, Is.EqualTo(wrappingFunctionMapEntry.DeminifiedMethodName));
+			Assert.That(stackFrameDeminification.DeminifiedStackFrame.SourcePosition, Is.EqualTo(SourcePosition.NotFound));
+		});
+		Assert.That(stackFrameDeminification.DeminifiedStackFrame.FilePath, Is.Null);
+	}
+
+	private static FunctionMapEntry CreateFunctionMapEntry(string deminifiedMethodName) => new(
+		Array.Empty<BindingInformation>(),
+		deminifiedMethodName,
+		default,
+		default);
+
+	private static SourceMap CreateSourceMap(List<MappingEntry>? parsedMappings = null) => new(
+		default,
+		default,
+		default,
+		default,
+		default,
+		parsedMappings ?? [],
+		default);
 }
