@@ -1,43 +1,41 @@
 ﻿using System.Text;
 using NUnit.Framework;
+using SourcemapTools.SourcemapParser.Internal;
 
-namespace SourcemapToolkit.SourcemapParser.UnitTests
+namespace SourcemapToolkit.SourcemapParser.UnitTests;
+
+public class Base64VlqEncoderUnitTests
 {
-
-	public class Base64VlqEncoderUnitTests
+	[Test]
+	public void Base64VlqEncoder_SmallValue_ListWithOnlyOneValue()
 	{
-		[Test]
-		public void Base64VlqEncoder_SmallValue_ListWithOnlyOneValue()
-		{
-			// Act
-			var result = new StringBuilder();
-			Base64VlqEncoder.Encode(result, 15);
+		// Act
+		var result = new StringBuilder();
+		Base64VlqEncoder.Encode(result, 15);
 
-			// Assert
-			Assert.AreEqual("e", result.ToString());
-		}
+		// Assert
+		Assert.That(result.ToString(), Is.EqualTo("e"));
+	}
 
-		[Test]
-		public void Base64VlqEncoder_LargeValue_ListWithOnlyMultipleValues()
-		{
-			// Act
-			var result = new StringBuilder();
-			Base64VlqEncoder.Encode(result, 701);
+	[Test]
+	public void Base64VlqEncoder_LargeValue_ListWithOnlyMultipleValues()
+	{
+		// Act
+		var result = new StringBuilder();
+		Base64VlqEncoder.Encode(result, 701);
 
-			// Assert
-			Assert.AreEqual("6rB", result.ToString());
-		}
+		// Assert
+		Assert.That(result.ToString(), Is.EqualTo("6rB"));
+	}
 
-		[Test]
-		public void Base64VlqEncoder_NegativeValue_ListWithCorrectValue()
-		{
-			// Act
-			var result = new StringBuilder();
-			Base64VlqEncoder.Encode(result, -15);
+	[Test]
+	public void Base64VlqEncoder_NegativeValue_ListWithCorrectValue()
+	{
+		// Act
+		var result = new StringBuilder();
+		Base64VlqEncoder.Encode(result, -15);
 
-			// Assert
-			Assert.AreEqual("f", result.ToString());
-		}
-
+		// Assert
+		Assert.That(result.ToString(), Is.EqualTo("f"));
 	}
 }
