@@ -2,13 +2,8 @@
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("SourcemapTools.UnitTests")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("SourcemapTools.UnitTests")]
-[assembly: AssemblyCopyright("Copyright ©  2021-2022")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
+[assembly: AssemblyCopyright("Copyright ©  2021-2023")]
 [assembly: ComVisible(false)]
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
